server/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
drf-yasg
python-dotenv
Pillow
django-filter
